requirements.txt

fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
alembic==1.13.2
celery==5.4.0
redis==5.0.7
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
pdfminer.six==20240706
pytesseract==0.3.13
Pillow==10.4.0
python-dotenv==1.0.1
boto3==1.34.154
minio==7.2.7
reportlab==4.2.2
fastapi-pagination==0.12.26
orjson==3.10.7
tenacity==9.0.0
tldextract==5.1.2
